# src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: PlayBall API
  description: |
    **PlayBall** - 스포츠 매칭 플랫폼 REST API 명세서

    ## 인증 방식
    카카오 OAuth 로그인 후 서버에서 JWT를 발급합니다.
    - `POST /api/auth/kakao` → 카카오 인가코드로 로그인 → accessToken + refreshToken 발급
    - 인증 필요 API: `Authorization: Bearer {accessToken}` 헤더 포함
    - 비로그인 허용 API: `security: []` 표기

    ## 공통 응답
    ```json
    { "success": true, "message": "성공", "data": { ... } }
    ```

    ## 에러 응답
    ```json
    { "success": false, "message": "에러 메시지", "data": null }
    ```

    ## 종목 코드
    `SOCCER` | `BASKETBALL` | `RUNNING` | `BADMINTON`

    ## 실력 코드
    `BEGINNER` | `INTERMEDIATE` | `ADVANCED`
  version: 1.1.0
  contact:
    name: PlayBall Team
servers:
  - url: http://localhost:8080
    description: 로컬 개발
  - url: https://api.playball.kr
    description: 운영
...
security:
  - BearerAuth: []

# ============================================================
paths:
...
  /api/matches:
    get:
      tags: [경기]
      summary: 경기 목록 조회
      description: |
        **위치 기반 조회 (latitude + longitude 제공 시)**
        - Haversine 공식으로 반경(radius km) 내 OPEN 경기를 거리순 반환
        - sportType 필터 적용 가능

        **일반 목록 조회 (위치 미제공 시)**
        - DELETED 제외한 전체 경기를 최신순, 페이지네이션으로 반환
      security: []
      parameters:
        - name: latitude
          in: query
          description: 검색 기준 위도 (longitude와 함께 제공 시 위치 기반 모드)
          schema:
            type: number
            format: double
          example: 37.6394
        - name: longitude
          in: query
          description: 검색 기준 경도
          schema:
            type: number
            format: double
          example: 127.0722
        - name: radius
          in: query
          description: 검색 반경 km (기본값 5.0, 위치 기반 모드에서만 사용)
          schema:
            type: number
            format: double
            default: 5.0
        - name: sportType
          in: query
          description: "종목 필터 (SOCCER | BASKETBALL | RUNNING | BADMINTON). 생략 시 전체"
          schema:
            type: string
        - name: page
          in: query
          description: 페이지 번호 (일반 목록 모드에서만 사용, 0부터 시작)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 페이지 크기 (일반 목록 모드에서만 사용)
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                locationBased:
                  summary: 위치 기반 (거리순)
                  value:
                    success: true
                    message: "경기 목록 조회 성공"
                    data:
                      - matchId: 2
                        title: "백석대 버스정류장 옆 농구"
                        sportType: "BASKETBALL"
                        matchDate: "2026-05-30T19:00:00"
                        locationName: "백석대 버스정류장 옆"
                        latitude: 37.6380
                        longitude: 127.0710
                        maxPlayers: 10
                        currentPlayers: 9
                        gender: null
                        ageRange: null
                        status: "OPEN"
                        updatedAt: null
                      - matchId: 1
                        title: "백석대 중앙운동장 축구"
                        sportType: "SOCCER"
                        matchDate: "2026-05-30T20:00:00"
                        locationName: "백석대 중앙운동장"
                        latitude: 37.6400
                        longitude: 127.0730
                        maxPlayers: 10
                        currentPlayers: 7
                        gender: "M"
                        ageRange: 20
                        status: "OPEN"
                        updatedAt: null
                paged:
                  summary: 일반 목록 (페이지네이션)
                  value:
                    success: true
                    message: "경기 목록 조회 성공"
                    data:
                      - matchId: 5
                        title: "강남 풋살"
                        sportType: "SOCCER"
                        matchDate: "2026-06-01T18:00:00"
                        locationName: "강남 풋살장"
                        latitude: 37.5172
                        longitude: 127.0473
                        maxPlayers: 10
                        currentPlayers: 3
                        gender: null
                        ageRange: null
                        status: "OPEN"
                        updatedAt: "2026-05-17T08:00:00"

    post:
      tags: [경기]
      summary: 경기 생성
      description: |
        경기를 생성합니다. 생성자는 자동으로 APPROVED 상태로 첫 번째 참가자가 됩니다.
        - `sportType`: `SOCCER` | `BASKETBALL` | `RUNNING` | `BADMINTON`
        - `skillLevel`: `BEGINNER` | `INTERMEDIATE` | `ADVANCED`
        - `matchDate`: 현재 시각 이후여야 합니다(`@FutureOrPresent`)
        - `entryFee`: 생략 또는 null 시 0으로 처리
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchCreateRequest'
            example:
              title: "백석대 컴공 축구 내기"
              sportType: "SOCCER"
              matchDate: "2026-05-30T19:00:00"
              locationName: "서울 목동운동장"
              latitude: 37.5283
              longitude: 126.8755
              address: "서울시 양천구 목동"
              maxPlayers: 10
              gender: "M"
              ageRange: 20
              skillLevel: "INTERMEDIATE"
              entryFee: 0
              description: "백석대 근처 축구하실 분 구합니다!"
      responses:
        '200':
          description: 경기 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "경기 생성 성공"
                data:
                  matchId: 1
                  title: "백석대 컴공 축구 내기"
                  sportType: "SOCCER"
                  matchDate: "2026-05-30T19:00:00"
                  locationName: "서울 목동운동장"
                  latitude: 37.5283
                  longitude: 126.8755
                  address: "서울시 양천구 목동"
                  maxPlayers: 10
                  currentPlayers: 1
                  gender: "M"
                  ageRange: 20
                  skillLevel: "INTERMEDIATE"
                  entryFee: 0
                  description: "백석대 근처 축구하실 분 구합니다!"
                  status: "OPEN"
        '400':
          description: 필수 항목 누락 또는 과거 날짜
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/matches/{matchId}:
    get:
      tags: [경기]
      summary: 경기 상세 조회
      description: |
        경기 정보와 APPROVED 상태인 참가자 목록을 함께 반환합니다.
        비로그인 사용자도 조회 가능합니다.
        삭제된 경기(DELETED)는 404를 반환합니다.
      security: []
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "경기 조회 성공"
                data:
                  match:
                    matchId: 1
                    title: "백석대 컴공 축구 내기"
                    sportType: "SOCCER"
                    matchDate: "2026-05-30T19:00:00"
                    locationName: "서울 목동운동장"
                    latitude: 37.5283
                    longitude: 126.8755
                    address: "서울시 양천구 목동"
                    maxPlayers: 10
                    currentPlayers: 3
                    gender: "M"
                    ageRange: 20
                    skillLevel: "INTERMEDIATE"
                    entryFee: 0
                    description: "백석대 근처 축구하실 분 구합니다!"
                    status: "OPEN"
                    createdAt: "2026-05-17T12:00:00"
                    updatedAt: "2026-05-17T12:00:00"
                  joinedMembers:
                    - memberId: 1
                      nickname: "홍길동"
                      profileImage: "https://k.kakaocdn.net/..."
                      gender: "M"
                      age: 22
                      skillLevel: "INTERMEDIATE"
                      preferredPosition: "ST"
                    - memberId: 2
                      nickname: "김철수"
                      profileImage: null
                      gender: "M"
                      age: 25
                      skillLevel: "BEGINNER"
                      preferredPosition: null
        '404':
          description: 존재하지 않거나 삭제된 경기
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value: { success: false, message: "존재하지 않는 경기입니다.", data: null }
                deleted:
                  value: { success: false, message: "이미 삭제된 경기입니다.", data: null }

    patch:
      tags: [경기]
      summary: 경기 수정 (주최자 전용)
      description: |
        변경할 필드만 포함하면 됩니다.
        - `maxPlayers`: 현재 참가 인원(`currentPlayers`)보다 작게 설정 불가
        - `matchDate`: 현재 시각 이후여야 합니다
        삭제된 경기는 수정 불가합니다.
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchUpdateRequest'
            example:
              title: "백석대 컴공 축구 내기 (인원 추가)"
              matchDate: "2026-05-30T19:00:00"
              maxPlayers: 12
              description: "인원 늘렸습니다!"
              entryFee: 0
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "경기 수정 성공"
                data:
                  matchId: 1
                  title: "백석대 컴공 축구 내기 (인원 추가)"
                  matchDate: "2026-05-30T19:00:00"
                  maxPlayers: 12
                  currentPlayers: 3
                  description: "인원 늘렸습니다!"
                  entryFee: 0
                  status: "OPEN"
                  updatedAt: "2026-05-17T16:00:00"
        '400':
          description: maxPlayers < currentPlayers 또는 과거 날짜
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { success: false, message: "잘못된 입력입니다.", data: null }
        '403':
          description: 주최자가 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { success: false, message: "접근 권한이 없습니다.", data: null }
        '404':
          $ref: '#/components/responses/MatchNotFound'

    delete:
      tags: [경기]
      summary: 경기 삭제 (주최자 전용)
      description: |
        경기를 DELETED 상태로 변경합니다(소프트 삭제).
        삭제 후 해당 경기는 목록 및 상세 조회에서 제외됩니다.
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "경기가 삭제되었습니다"
                data: null
        '403':
          description: 주최자가 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/MatchNotFound'

  # ──────────────────────────────────────────────
  # 매칭 (참가·취소·랜덤)
  # ──────────────────────────────────────────────
  /api/matches/{matchId}/join:
    post:
      tags: [매칭]
      summary: 경기 참가
      description: |
        경기에 참가를 신청합니다. 참가 즉시 APPROVED 상태가 됩니다.

        **검증 순서:**
        1. 경기 존재 여부 확인 → 없으면 404
        2. DELETED 상태 확인 → 삭제된 경기면 404
        3. 인원 초과 확인 → 꽉 찼으면 409
        4. 이미 참가 여부 확인(PENDING or APPROVED) → 409
        5. 참가 등록 + currentPlayers 증가

        인원이 maxPlayers에 달하면 status가 `CLOSED`로 자동 변경됩니다.
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 참가 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "경기 참가 성공"
                data:
                  matchId: 1
                  currentPlayers: 4
                  maxPlayers: 10
                  status: "OPEN"
        '404':
          description: 존재하지 않거나 삭제된 경기
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value: { success: false, message: "존재하지 않는 경기입니다.", data: null }
                deleted:
                  value: { success: false, message: "이미 삭제된 경기입니다.", data: null }
        '409':
          description: 인원 초과 또는 이미 참가
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                full:
                  summary: 인원 초과
                  value: { success: false, message: "경기 인원이 꽉 찼습니다.", data: null }
                alreadyJoined:
                  summary: 이미 참가
                  value: { success: false, message: "이미 참가한 경기입니다.", data: null }
                conflict:
                  summary: 동시 참가 충돌 (낙관적 락)
                  value: { success: false, message: "다른 사용자가 동시에 참가 중입니다. 다시 시도해주세요.", data: null }

    delete:
      tags: [매칭]
      summary: 경기 참가 취소
      description: |
        내가 APPROVED 상태로 참가 중인 경기에서 나갑니다.

        **검증:**
        - DELETED 경기면 404
        - APPROVED 상태 참가 기록이 없으면 400 (`NOT_JOINED`)
        - 성공 시 참가 기록을 CANCELLED로 변경 + currentPlayers 감소
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 취소 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "경기 참가 취소 성공"
                data:
                  matchId: 1
                  currentPlayers: 3
                  maxPlayers: 10
                  status: "OPEN"
        '400':
          description: 참가하지 않은 경기
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { success: false, message: "참가하지 않은 경기입니다.", data: null }
        '404':
          $ref: '#/components/responses/MatchNotFound'
  /api/matches/random:
    post:
      tags: [매칭]
      summary: 랜덤 매칭
      description: |
        조건에 맞는 경기를 랜덤으로 1개 추천합니다.

        **필터 조건 (모두 선택):**
        - 위치 기반 반경 내
        - OPEN 상태 + currentPlayers < maxPlayers
        - 본인이 이미 참가(PENDING or APPROVED)한 경기 제외
        - sportType, date, maxFee, skillLevel, gender, ageRange 필터

        조건에 맞는 경기가 없으면 404를 반환합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RandomMatchRequest'
            example:
              latitude: 37.6394
              longitude: 127.0722
              radius: 5.0
              sportType: "SOCCER"
              date: "2026-05-30"
              maxFee: 10000
              skillLevel: "INTERMEDIATE"
              gender: "M"
              ageRange: 20
      responses:
        '200':
          description: 매칭 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "랜덤 매칭 성공"
                data:
                  matchId: 3
                  title: "백석대 축구 내기"
                  sportType: "SOCCER"
                  matchDate: "2026-05-30T19:00:00"
                  locationName: "서울 목동운동장"
                  entryFee: 0
                  currentPlayers: 7
                  maxPlayers: 10
                  distance: 2.3
        '404':
          description: 조건에 맞는 경기 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { success: false, message: "존재하지 않는 경기입니다.", data: null }

  # ──────────────────────────────────────────────
  # 알림
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "카카오 로그인 후 발급받은 accessToken"
  responses:
    Unauthorized:
      description: 인증 실패 (토큰 없음 또는 만료)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "인증이 필요합니다."
            data: null
    MatchNotFound:
      description: 존재하지 않거나 삭제된 경기
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value: { success: false, message: "존재하지 않는 경기입니다.", data: null }
            deleted:
              value: { success: false, message: "이미 삭제된 경기입니다.", data: null }
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        data:
          nullable: true
          example: null
...
    MatchCreateRequest:
      type: object
      required: [title, sportType, matchDate, latitude, longitude, maxPlayers]
      properties:
        title:
          type: string
        sportType:
          type: string
          enum: [SOCCER, BASKETBALL, RUNNING, BADMINTON]
        matchDate:
          type: string
          format: date-time
          description: "현재 시각 이후여야 함 (@FutureOrPresent)"
        locationName:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address:
          type: string
        maxPlayers:
          type: integer
          minimum: 1
        gender:
          type: string
          enum: [M, F]
          nullable: true
          description: "null이면 무관"
        ageRange:
          type: integer
          nullable: true
          description: "연령대 (예: 20 = 20대). null이면 무관"
        skillLevel:
          type: string
          enum: [BEGINNER, INTERMEDIATE, ADVANCED]
          nullable: true
        entryFee:
          type: integer
          minimum: 0
          nullable: true
          description: "참가비. null이면 0으로 처리"
        description:
          type: string
          nullable: true
    MatchUpdateRequest:
      type: object
      description: 변경할 필드만 포함. null이면 해당 필드 유지.
      properties:
        title:
          type: string
          nullable: true
        matchDate:
          type: string
          format: date-time
          nullable: true
          description: "현재 시각 이후여야 함 (@FutureOrPresent)"
        maxPlayers:
          type: integer
          minimum: 1
          nullable: true
          description: "currentPlayers 이상이어야 함"
        entryFee:
          type: integer
          minimum: 0
          nullable: true
        description:
          type: string
          nullable: true
    RandomMatchRequest:
      type: object
      required: [latitude, longitude, sportType]
      properties:
        latitude:
          type: number
          format: double
          description: "현재 위치 위도"
        longitude:
          type: number
          format: double
          description: "현재 위치 경도"
        address:
          type: string
          nullable: true
          description: "현재 주소 (선택)"
        radius:
          type: number
          format: double
          default: 5.0
          description: "검색 반경 km"
        sportType:
          type: string
          enum: [SOCCER, BASKETBALL, RUNNING, BADMINTON]
        date:
          type: string
          format: date
          nullable: true
          description: "희망 경기 날짜 (생략 시 날짜 무관)"
        preferredPosition:
          type: string
          nullable: true
          description: "선호 포지션 (선택)"
        maxFee:
          type: integer
          nullable: true
          description: "최대 참가비 필터 (이 금액 이하만 조회)"
        skillLevel:
          type: string
          enum: [BEGINNER, INTERMEDIATE, ADVANCED]
          nullable: true
        gender:
          type: string
          enum: [M, F]
          nullable: true
        ageRange:
          type: integer
          nullable: true
...